requirements-dev.txt
# Testing
pytest
pytest-cov
httpx

# Linting
flake8

# Type checking
mypy
pandas-stubs